+++ src/main/kotlin/step4/racingcar/Main.kt
@@ -1,5 +1,6 @@
 package step4.racingcar
 
+import step4.racingcar.controller.CarRacingController
 import step4.racingcar.domain.Car
 import step4.racingcar.domain.CarRacing
 import step4.racingcar.domain.RandomNumber
@@ -7,11 +8,5 @@ import step4.racingcar.view.InputView
 import step4.racingcar.view.ResultView
 
 fun main() {
-
-    val racers = InputView.racers()
-    val movementCount = InputView.movementCount()
-
-    val racingGame = CarRacing(racers.map { Car(it) }, movementCount)
-    racingGame.execute(RandomNumber)
-    ResultView.view(racingGame)
+    CarRacingController.play()
 }
